First implementation. Next problem to solve: see comment.
For InputTextureSound there's a set_position method called for each frame.
The issue is that, as of now, the image is recreated from scratch -
and OpenGL doesn't like that:
Cannot set mip levels on a texture that already has storage allocated.
To do so, destroy() the texture and then create() and setMipLevels()