Spectrum part is more or less done. See comment.
I've changed the way of preparing the image of the spectrum, manually filling the magnitudes from FFT.
Also I got rid of Hann filtering, as scaling the results to dB scale was broken after it - and it wasn't
introducing any real value in this use case.
Now, I want to take a look into timing and syncing with the audio file.
Also, dealing with  the file end, trailing with zeroes etc. has to be done.